Create Configuration.cs for having information about clients, resources etc.
ApiScope - represents what the client application can use, that is, an identifier for resources

In IdentityServer, areas are represented by resources
these can be identity resources and api resources.

IdenityResource allows you to model a scope that will allow a client application to view a subset of claims about a user

Client (clients) - identity servery needs to know which client applications are allowed to use it.